backend/requirements.txt
Flask==2.3.0
Flask-Cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==2.3.0
PyJWT==2.8.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
eventlet==0.33.3
